refactor(queries): extract reusable i18n query fragments
