+++ studio/lib/queries/utils/i18n.ts
@@ -1,0 +1,5 @@
+import { groq } from "next-sanity";
+
+export function translatedFieldFragment(fieldName: string) {
+  return groq`${fieldName}[_key == $language][0].value`;
+}
